requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0
python-docx>=0.8.11
PyPDF2>=3.0.0
plotly>=5.17.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3
openai>=1.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
urllib3>=2.1.0
